pyproject.toml
[project]
name = "adk-claw"
version = "0.1.0"
description = "OpenClaw-inspired autonomous AI agent using Google ADK + Gemini"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0",
    "google-genai",
    "python-telegram-bot>=21.0",
    "pydantic-settings",
    "python-dotenv",
    "rank-bm25",
    "httpx",
    "numpy",
    "sqlite-vec",
    "langfuse",
    "openinference-instrumentation-google-adk",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_claw"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
